First renamed errBot.py to core.py
Rationale: bad spelling for a python module, already in an 'errbot'
module anyway.
