+++ errbot/backends/null.py
@@ -3,7 +3,7 @@ from time import sleep
 from errbot.backends.base import ONLINE
 
 from errbot.backends.test import TestPerson
-from errbot.errBot import ErrBot
+from errbot.core import ErrBot
 
 
 # Can't use __name__ because of Yapsy
