+++ errbot/backends/telegram_messenger.py
@@ -3,7 +3,7 @@ import sys
 
 from errbot import PY2
 from errbot.backends.base import RoomError, Identifier, Person, RoomOccupant, ONLINE, Room
-from errbot.errBot import ErrBot
+from errbot.core import ErrBot
 from errbot.rendering import text
 from errbot.rendering.ansiext import enable_format, TEXT_CHRS
