+++ errbot/core.py
