+++ tests/backend_tests/slack_tests.py
@@ -6,7 +6,7 @@ from tempfile import mkdtemp
 
 import mock
 
-from errbot.errBot import bot_config_defaults
+from errbot.core import bot_config_defaults
 
 log = logging.getLogger(__name__)
