+++ tests/base_backend_tests.py
@@ -6,13 +6,13 @@ from os.path import sep
 
 import pytest
 
-from errbot.errBot import bot_config_defaults
+from errbot.core import bot_config_defaults
 
 import os  # noqa
 import re  # noqa
 from collections import OrderedDict
 from queue import Queue, Empty  # noqa
-from errbot.errBot import ErrBot
+from errbot.core import ErrBot
 from errbot.backends.base import Message, Room, Identifier, ONLINE
 from errbot.backends.test import TestPerson, TestOccupant, TestRoom, ShallowConfig
 from errbot import botcmd, re_botcmd, arg_botcmd, templating  # noqa
