+++ tests/utils_tests.py
@@ -3,7 +3,7 @@ from datetime import timedelta
 import pytest
 
 from errbot.backends.test import ShallowConfig
-from errbot.errBot import bot_config_defaults
+from errbot.core import bot_config_defaults
 from errbot.main import CORE_STORAGE
 from errbot.specific_plugin_manager import SpecificPluginManager
 from errbot.storage.base import StoragePluginBase
